+++ common/src/main/java/me/syldium/thimble/common/command/game/JoinCommand.java
@@ -1,6 +1,8 @@
 package me.syldium.thimble.common.command.game;
 
+import me.syldium.thimble.api.arena.ThimbleArena;
 import me.syldium.thimble.api.arena.ThimbleGame;
+import me.syldium.thimble.api.service.GameService;
 import me.syldium.thimble.common.ThimblePlugin;
 import me.syldium.thimble.common.command.abstraction.ChildCommand;
 import me.syldium.thimble.common.command.abstraction.CommandException;
@@ -13,18 +15,27 @@ import me.syldium.thimble.common.game.Arena;
 import me.syldium.thimble.common.player.MessageKey;
 import me.syldium.thimble.common.player.Player;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.Optional;
 
 public class JoinCommand extends ChildCommand.One<Arena> {
 
     public JoinCommand() {
-        super("join", Arguments.arena(), MessageKey.HELP_JOIN, Permission.player("join"));
+        super("join", Arguments.arena().optional(), MessageKey.HELP_JOIN, Permission.player("join"));
         this.commandGuard = CommandGuard.EXCEPT_NOT_IN_GAME;
     }
 
     @Override
-    public @NotNull CommandResult execute(@NotNull ThimblePlugin plugin, @NotNull Sender sender, @NotNull Arena arena) throws CommandException {
+    public @NotNull CommandResult execute(@NotNull ThimblePlugin plugin, @NotNull Sender sender, @Nullable Arena arena) throws CommandException {
+        if (arena == null) {
+            Optional<ThimbleArena> arenaOpt = plugin.getGameService().findAvailableArena(GameService.ArenaSelection.MOST_FILLED, 1);
+            if (arenaOpt.isPresent()) {
+                arena = (Arena) arenaOpt.get();
+            } else {
+                throw new CommandException(MessageKey.FEEDBACK_GAME_FULL_AUTO);
+            }
+        }
         if (!arena.isSetup()) {
             throw new CommandException(MessageKey.FEEDBACK_ARENA_NOT_CONFIGURED);
         }